Add SegmentedControlWithoutStyles
